+++ app/components/admin/budgets/duration_in_words_component.html.erb
@@ -1,0 +1,1 @@
+<%= duration -%>
