+++ app/components/admin/duration_component.html.erb
@@ -1,0 +1,1 @@
+<%= dates -%>
